us12_finaltask_created_test_folder_to_separate_coomon_conditions_and_tests
